backend/ai_vice_backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.2
openai==2.2.0
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.1
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
wsproto==1.2.0
